Initializes the princip of Color Themes
Creates The class Design and the DefaultDesign (inherits from Design )class, which is the default value of the Class ColorTheme (on which will be accessed by the windows and elements).

The colors of the defaultDesign are still random. This must be corrected and also new inhereting classes in other themes.

relates to #5